+++ docs/astro.config.ts
@@ -1,7 +1,22 @@
 import { defineConfig } from 'astro/config';
-import starWarp from '@inox-tools/star-warp';
 import starlight from '@astrojs/starlight';
+import type { StarlightConfig } from '@astrojs/starlight/types';
+import vercel from '@astrojs/vercel';
+import starWarp from '@inox-tools/star-warp';
 import starlightLinksValidator from 'starlight-links-validator';
+
+const badge = {
+	new: {
+		text: 'NEW',
+		variant: 'success',
+	},
+	updated: {
+		text: 'UPDATED',
+		variant: 'default',
+	},
+} satisfies Record<string, NonNullable<NonNullable<StarlightConfig['sidebar']>[number]['badge']>>;
+
+process.env.ASTRO_PROJECT_ROOT = new URL('../', import.meta.url).toString();
 
 const SITE =
 	process.env.VERCEL_ENV !== 'production' && process.env.VERCEL_URL
@@ -10,6 +25,9 @@ const SITE =
 
 // https://astro.build/config
 export default defineConfig({
+	adapter: vercel({
+		skewProtection: true,
+	}),
 	site: SITE,
 	trailingSlash: 'never',
 	integrations: [
@@ -24,6 +42,8 @@ export default defineConfig({
 			components: {
 				Head: './src/components/Head.astro',
 				PageTitle: './src/components/PageTitle.astro',
+				Sidebar: './src/components/Sidebar.astro',
+				MarkdownContent: './src/components/MarkdownContent.astro',
 			},
 			sidebar: [
 				{
@@ -53,39 +73,53 @@ export default defineConfig({
 						{
 							label: 'Request State',
 							link: '/request-state',
-							badge: {
-								text: 'NEW',
-								variant: 'success',
-							},
 						},
 						{
 							label: 'Request Nanostores',
 							link: '/request-nanostores',
-							badge: {
-								text: 'NEW',
-								variant: 'success',
+						},
+						{
+							label: 'Cut Short',
+							link: '/cut-short',
+						},
+						{
+							label: 'Portal Gun',
+							link: '/portal-gun',
+						},
+						{
+							label: 'Content Utilities',
+							collapsed: false,
+							autogenerate: {
+								directory: 'content-utils',
 							},
 						},
 					],
 				},
 				{
-					label: 'Modular Station',
+					label: 'Tools for Authors',
 					collapsed: false,
-					autogenerate: {
-						directory: 'modular-station',
-					},
-				},
-				{
-					label: 'Content Utilities',
-					collapsed: false,
-					autogenerate: {
-						directory: 'content-utils',
-					},
-				},
-				{
-					label: 'Inline Module',
-					collapsed: false,
-					autogenerate: { directory: 'inline-mod' },
+					items: [
+						{
+							label: 'Astro Integration Kit',
+							link: 'https://astro-integration-kit.netlify.app',
+						},
+						{
+							label: 'Astro Tests',
+							link: '/astro-tests',
+						},
+						{
+							label: 'Modular Station',
+							collapsed: false,
+							autogenerate: {
+								directory: 'modular-station',
+							},
+						},
+						{
+							label: 'Inline Module',
+							collapsed: true,
+							autogenerate: { directory: 'inline-mod' },
+						},
+					],
 				},
 			],
 			plugins: [
@@ -100,5 +134,8 @@ export default defineConfig({
 		'/content-utils': '/content-utils/integration',
 		'/content-utils/git-time': '/content-utils/git',
 		'/modular-station': '/modular-station/api',
+	},
+	image: {
+		domains: ['mermaid.ink'],
 	},
 });