+++ packages/request-nanostores/e2e/fixture/basic/astro.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from 'astro/config';
+import node from '@astrojs/node';
+import requestNanostores from '@inox-tools/request-nanostores';
+
+import preact from '@astrojs/preact';
+
+export default defineConfig({
+	output: 'server',
+	adapter: node({ mode: 'standalone' }),
+	integrations: [requestNanostores(), preact()],
+	devToolbar: { enabled: false },
+});