+++ packages/sitemap-ext/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @inox-tools/sitemap-ext
 
+## 0.4.1
+
+### Patch Changes
+
+- 0dcdeaa: Bump Astro Integration Kit
+- Updated dependencies [0dcdeaa]
+  - @inox-tools/aik-route-config@0.8.1
+
+## 0.4.0
+
+### Minor Changes
+
+- Minimal bump to sync inter-dependencies to Astro 5 and Vite 6
+
+### Patch Changes
+
+- Updated dependencies
+  - @inox-tools/aik-route-config@0.8.0
+
+## 0.3.6
+
+### Patch Changes
+
+- @inox-tools/aik-route-config@0.7.5
+
+## 0.3.6-beta.0
+
+### Patch Changes
+
+- @inox-tools/aik-route-config@0.7.5-beta.0
+
+## 0.3.5
+
+### Patch Changes
+
+- @inox-tools/aik-route-config@0.7.4
+
+## 0.3.4
+
+### Patch Changes
+
+- @inox-tools/aik-route-config@0.7.3
+
 ## 0.3.3
 
 ### Patch Changes